Fix and improve the minimum order quantity, attrition and pricing.
There were a number of problems preventing the existing code from working.

  1. Changes at the JLC website relocating APIs
  2. The CORS proxy that was in use was no longer able to request from JLC
  3. The way the JLC search works requires several different searches to be sure to find components

The loading fee for parts (as of writing, $3.00 USD for extended parts in Economical PCBA, and $1.50 for both basic/preferred and extended parts in Standard PCBA) has been added to the display and incorporated in the total and amortised total per-piece prices.

A note has been added to indicate how the total is calculated to make it clear how the attrition factors into the equation.

HTML djustments have been made for aesthetic reasons.

See dougy83/jlcparts#7